+++ cypress/e2e/feed.cy.ts
@@ -170,6 +170,7 @@ sizes.forEach((size) => {
           activeCollectionTitle: 'Kent C. Dodds Blog',
           activeCollectionFilter: 'all',
           activeCollectionGrouping: 'none',
+          activeCollectionSortBy: 'newestFirst',
         });
         expect(response.statusCode).to.eq(200);
       });
@@ -225,7 +226,9 @@ sizes.forEach((size) => {
       });
       cy.intercept({
         method: 'GET',
-        url: getApiPath('/collections/2/items?pageIndex=0&filter=all'),
+        url: getApiPath(
+          '/collections/2/items?pageIndex=0&filter=all&grouping=none&sortBy=newestFirst'
+        ),
       }).as('apiGetItems');
       cy.intercept({
         method: 'GET',
@@ -301,7 +304,9 @@ sizes.forEach((size) => {
       it('from collection header', () => {
         cy.intercept({
           method: 'GET',
-          url: getApiPath('/collections/2/items?pageIndex=0&filter=all'),
+          url: getApiPath(
+            '/collections/2/items?pageIndex=0&filter=all&grouping=none&sortBy=newestFirst'
+          ),
         }).as('apiGetItems');
 
         cy.intercept({
@@ -358,7 +363,9 @@ sizes.forEach((size) => {
 
         cy.intercept({
           method: 'GET',
-          url: getApiPath('/collections/1/items?pageIndex=0&filter=all'),
+          url: getApiPath(
+            '/collections/1/items?pageIndex=0&filter=all&grouping=none&sortBy=newestFirst'
+          ),
         }).as('apiGetHomeItems');
 
         cy.addFeedByApi({
@@ -404,7 +411,9 @@ sizes.forEach((size) => {
         }).as('apiMarkAsRead');
         cy.intercept({
           method: 'GET',
-          url: getApiPath('/collections/1/items?pageIndex=0&filter=all'),
+          url: getApiPath(
+            '/collections/1/items?pageIndex=0&filter=all&grouping=none&sortBy=newestFirst'
+          ),
         }).as('apiGetHomeItems');
 
         cy.addFeedByApi({
@@ -474,7 +483,9 @@ sizes.forEach((size) => {
           it(`to ${layout}`, () => {
             cy.intercept({
               method: 'GET',
-              url: getApiPath('/collections/2/items?pageIndex=0&filter=all'),
+              url: getApiPath(
+                '/collections/2/items?pageIndex=0&filter=all&grouping=none&sortBy=newestFirst'
+              ),
             }).as('apiGetItems');
 
             cy.intercept({
@@ -536,7 +547,9 @@ sizes.forEach((size) => {
         }).as('apiDeleteCollection');
         cy.intercept({
           method: 'GET',
-          url: getApiPath(`/collections/1/items?pageIndex=0&filter=all`),
+          url: getApiPath(
+            `/collections/1/items?pageIndex=0&filter=all&grouping=none&sortBy=newestFirst`
+          ),
         }).as('apiGetHomeItems');
 
         cy.getBySel('confirmDelete').click();
@@ -578,7 +591,9 @@ sizes.forEach((size) => {
         }).as('apiDeleteCollection');
         cy.intercept({
           method: 'GET',
-          url: getApiPath(`/collections/1/items?pageIndex=0&filter=all`),
+          url: getApiPath(
+            `/collections/1/items?pageIndex=0&filter=all&grouping=none&sortBy=newestFirst`
+          ),
         }).as('apiGetHomeItems');
         cy.intercept({
           method: 'PUT',
@@ -603,6 +618,7 @@ sizes.forEach((size) => {
             activeCollectionFilter: 'all',
             activeCollectionGrouping: 'none',
             activeCollectionLayout: 'card',
+            activeCollectionSortBy: 'newestFirst',
           });
           expect(response.statusCode).to.eq(200);
         });
@@ -641,7 +657,9 @@ sizes.forEach((size) => {
         }).as('apiDeleteCollection');
         cy.intercept({
           method: 'GET',
-          url: getApiPath(`/collections/1/items?pageIndex=0&filter=all`),
+          url: getApiPath(
+            `/collections/1/items?pageIndex=0&filter=all&grouping=none&sortBy=newestFirst`
+          ),
         }).as('apiGetHomeItems');
 
         cy.getBySel('confirmDelete').click();
@@ -700,11 +718,15 @@ sizes.forEach((size) => {
 
         cy.intercept({
           method: 'GET',
-          url: getApiPath(`/collections/2/items?pageIndex=0&filter=unread`),
+          url: getApiPath(
+            `/collections/2/items?pageIndex=0&filter=unread&grouping=none&sortBy=newestFirst`
+          ),
         }).as('apiGetItems2');
         cy.intercept({
           method: 'GET',
-          url: getApiPath(`/collections/3/items?pageIndex=0&filter=read`),
+          url: getApiPath(
+            `/collections/3/items?pageIndex=0&filter=read&grouping=none&sortBy=newestFirst`
+          ),
         }).as('apiGetItems3');
 
         cy.visit('/');
@@ -720,6 +742,167 @@ sizes.forEach((size) => {
         cy.wait('@apiGetItems3').then(({ response }) => {
           expect(response.body).to.have.length(0);
         });
+      });
+    });
+
+    describe('grouping', () => {
+      beforeEach(() => {
+        cy.putCollectionPreferencesByApi({
+          collectionId: 1,
+          preferences: { layout: 'list' },
+        });
+
+        cy.addFeedByApi({
+          title: 'Kent C. Dodds Blog',
+          url: getFeedUrl('kentcdodds.xml'),
+          icon: 'Code',
+          refreshInterval: 120,
+        });
+
+        cy.addFeedByApi({
+          title: 'fettblog.eu',
+          url: getFeedUrl('fettblog.xml'),
+          icon: 'Code',
+          refreshInterval: 120,
+        });
+      });
+
+      it('by feed', () => {
+        cy.intercept({
+          method: 'GET',
+          url: getApiPath(
+            `/collections/1/items?pageIndex=0&filter=all&grouping=none&sortBy=newestFirst`
+          ),
+        }).as('apiGetItemsWithoutGrouping');
+
+        cy.intercept({
+          method: 'GET',
+          url: getApiPath(
+            `/collections/1/items?pageIndex=0&filter=all&grouping=feed&sortBy=newestFirst`
+          ),
+        }).as('apiGetItemsWithGrouping');
+
+        cy.visit('/');
+        cy.wait('@apiGetItemsWithoutGrouping');
+
+        cy.clickCollectionHeaderMenuAction('grouping-feed');
+        cy.wait('@apiGetItemsWithGrouping');
+
+        cy.getBySel('groupHeader-fettblog.eu').should('be.visible');
+        cy.getBySel('groupHeader-Kent C. Dodds Blog').should('be.visible');
+      });
+
+      describe('by date', () => {
+        beforeEach(() => {
+          cy.intercept({
+            method: 'GET',
+            url: getApiPath(
+              `/collections/1/items?pageIndex=0&filter=all&grouping=none&sortBy=newestFirst`
+            ),
+          }).as('apiGetItemsWithoutGrouping');
+
+          cy.intercept({
+            method: 'GET',
+            url: getApiPath(
+              `/collections/1/items?pageIndex=0&filter=all&grouping=date&sortBy=newestFirst`
+            ),
+          }).as('apiGetItemsWithGrouping');
+        });
+
+        it('Today', () => {
+          cy.clock(new Date(2022, 4, 11), ['Date']);
+          cy.visit('/');
+          cy.wait('@apiGetItemsWithoutGrouping');
+
+          cy.clickCollectionHeaderMenuAction('grouping-date');
+          cy.wait('@apiGetItemsWithGrouping');
+
+          cy.getBySel('groupHeader-Today').should('be.visible');
+          cy.getBySel('groupHeader-This week').should('be.visible');
+          cy.getBySel('groupHeader-Over a month ago').should('be.visible');
+        });
+
+        it('Yesterday', () => {
+          cy.clock(new Date(2022, 4, 12), ['Date']);
+          cy.visit('/');
+          cy.wait('@apiGetItemsWithoutGrouping');
+
+          cy.clickCollectionHeaderMenuAction('grouping-date');
+          cy.wait('@apiGetItemsWithGrouping');
+
+          cy.getBySel('groupHeader-Yesterday').should('be.visible');
+          cy.getBySel('groupHeader-This week').should('be.visible');
+          cy.getBySel('groupHeader-Over a month ago').should('be.visible');
+        });
+      });
+    });
+
+    describe('sorting', () => {
+      beforeEach(() => {
+        cy.putCollectionPreferencesByApi({
+          collectionId: 1,
+          preferences: { layout: 'list' },
+        });
+
+        cy.addFeedByApi({
+          title: 'Kent C. Dodds Blog',
+          url: getFeedUrl('kentcdodds.xml'),
+          icon: 'Code',
+          refreshInterval: 120,
+        });
+      });
+
+      it('oldest first', () => {
+        cy.intercept({
+          method: 'GET',
+          url: getApiPath(
+            `/collections/1/items?pageIndex=0&filter=all&grouping=none&sortBy=newestFirst`
+          ),
+        }).as('apiGetItems');
+
+        cy.intercept({
+          method: 'GET',
+          url: getApiPath(
+            `/collections/1/items?pageIndex=0&filter=all&grouping=none&sortBy=oldestFirst`
+          ),
+        }).as('apiGetItemsOldestFirst');
+
+        cy.visit('/');
+        cy.wait('@apiGetItems');
+
+        cy.clickCollectionHeaderMenuAction('sortBy-oldestFirst');
+        cy.wait('@apiGetItemsOldestFirst');
+
+        cy.getBySel('collectionItemList')
+          .within(() => {
+            cy.get('[data-index=0]')
+              .within(() => {
+                cy.getBySelVisible('title').should(
+                  'have.text',
+                  'How I help you build better websites'
+                );
+              })
+              .should('exist');
+
+            cy.get('[data-index=1]')
+              .within(() => {
+                cy.getBySelVisible('title').should(
+                  'have.text',
+                  'Why I Love Remix'
+                );
+              })
+              .should('exist');
+
+            cy.get('[data-index=2]')
+              .within(() => {
+                cy.getBySelVisible('title').should(
+                  'have.text',
+                  "Remix: The Yang to React's Yin"
+                );
+              })
+              .should('exist');
+          })
+          .should('exist');
       });
     });
   });