+++ cypress/support/commands.ts
@@ -30,15 +30,15 @@ import { getApiPath } from '../e2e/utils';
 //
 
 Cypress.Commands.add('getBySel', (selector, ...args) => {
-  return cy.get(`[data-test=${selector}]`, ...args);
+  return cy.get(`[data-test="${selector}"]`, ...args);
 });
 
 Cypress.Commands.add('getBySelVisible', (selector, ...args) => {
-  return cy.get(`[data-test=${selector}]:visible`, ...args);
+  return cy.get(`[data-test="${selector}"]:visible`, ...args);
 });
 
 Cypress.Commands.add('getBySelLike', (selector, ...args) => {
-  return cy.get(`[data-test*=${selector}]`, ...args);
+  return cy.get(`[data-test*="${selector}"]`, ...args);
 });
 
 Cypress.Commands.add('openDrawerIfExists', () => {
@@ -131,7 +131,9 @@ Cypress.Commands.add('clickGoBackIfExists', () => {
 Cypress.Commands.add('changeActiveCollection', (id: string) => {
   cy.intercept({
     method: 'GET',
-    url: getApiPath(`/collections/${id}/items?pageIndex=0&filter=all`),
+    url: getApiPath(
+      `/collections/${id}/items?pageIndex=0&filter=all&grouping=none&sortBy=newestFirst`
+    ),
   }).as('apiGetItems');
 
   cy.visit('/');