+++ packages/backend/db/collectionItems/sql.ts
@@ -1,5 +1,10 @@
 import { sql } from 'slonik';
-import { CollectionFilter, ID } from '@orpington-news/shared';
+import {
+  CollectionFilter,
+  CollectionGrouping,
+  CollectionSortBy,
+  ID,
+} from '@orpington-news/shared';
 import { getCollectionChildrenIds } from '@db/collections';
 import { TRUE } from '@utils';
 import { DBCollectionItem, DBCollectionItemWithoutText } from './types';
@@ -72,12 +77,16 @@ export interface GetCollectionItemsArgs {
   userId: ID;
   collectionId: ID;
   filter: CollectionFilter;
+  grouping: CollectionGrouping;
+  sortBy: CollectionSortBy;
 }
 
 export const getCollectionItems = ({
   userId,
   collectionId,
   filter,
+  grouping,
+  sortBy,
 }: GetCollectionItemsArgs) => {
   const showFilter =
     filter === 'all'
@@ -85,6 +94,13 @@ export const getCollectionItems = ({
       : filter === 'unread'
       ? sql`collection_items.date_read IS NULL`
       : sql`collection_items.date_read IS NOT NULL`;
+
+  const sort = sortBy === 'newestFirst' ? sql`DESC` : sql`ASC`;
+
+  const orderBy =
+    grouping === 'feed'
+      ? sql`lower(collection_title) ASC, date_published ${sort}`
+      : sql`date_published ${sort}`;
 
   return sql.type(DBCollectionItemWithoutText)`
 SELECT
@@ -118,7 +134,7 @@ WHERE
   AND
   ${showFilter}
 ORDER BY
-  date_published DESC
+  ${orderBy}
 `;
 };