+++ packages/frontend/components/collection/header/CollectionHeader.tsx
@@ -23,8 +23,16 @@ import {
   CollectionLayout,
   CollectionFilter,
   ActiveCollection,
+  CollectionGrouping,
+  CollectionSortBy,
+  CollectionPreferences,
 } from '@orpington-news/shared';
-import { CollectionLayoutName, CollectionFilterName } from '../types';
+import {
+  CollectionLayoutName,
+  CollectionFilterName,
+  CollectionGroupingName,
+  CollectionSortByName,
+} from '../types';
 
 export type MenuAction = 'refresh' | 'markAsRead';
 
@@ -34,9 +42,8 @@ export interface CollectionHeaderProps {
   isRefreshing?: boolean;
 
   onHamburgerClicked?: () => void;
-  onChangeLayout?: (layout: CollectionLayout) => void;
   onMenuActionClicked?: (action: MenuAction) => void;
-  onShowFilterChanged?: (showFilter: CollectionFilter) => void;
+  onPreferenceChanged?: (preferences: CollectionPreferences) => void;
 }
 
 export const CollectionHeader: React.FC<CollectionHeaderProps> = (props) => {
@@ -45,9 +52,8 @@ export const CollectionHeader: React.FC<CollectionHeaderProps> = (props) => {
     menuButtonRef,
     isRefreshing = false,
     onHamburgerClicked,
-    onChangeLayout,
     onMenuActionClicked,
-    onShowFilterChanged,
+    onPreferenceChanged,
   } = props;
 
   const isLoading = collection === undefined;
@@ -104,7 +110,7 @@ export const CollectionHeader: React.FC<CollectionHeaderProps> = (props) => {
                       <MenuItemOption
                         key={layout}
                         value={layout}
-                        onClick={() => onChangeLayout?.(layout)}
+                        onClick={() => onPreferenceChanged?.({ layout })}
                         data-test={`layout-${layout}`}
                       >
                         {CollectionLayoutName[layout]}
@@ -127,14 +133,19 @@ export const CollectionHeader: React.FC<CollectionHeaderProps> = (props) => {
                 tabIndex={0}
                 data-test="menuButton"
               />
-              <MenuList data-focus-visible-disabled data-test="menuList">
+              <MenuList
+                data-focus-visible-disabled
+                data-test="menuList"
+                zIndex="docked"
+              >
                 <MenuItem
                   icon={<IoCheckmarkDone />}
                   onClick={() => onMenuActionClicked?.('markAsRead')}
                   data-test="markAsRead"
                 >
-                  Mark as read
+                  Mark all as read
                 </MenuItem>
+
                 <MenuDivider />
 
                 <MenuOptionGroup
@@ -142,14 +153,52 @@ export const CollectionHeader: React.FC<CollectionHeaderProps> = (props) => {
                   title="Show"
                   type="radio"
                 >
-                  {CollectionFilter.options.map((showFilter) => (
+                  {CollectionFilter.options.map((filter) => (
                     <MenuItemOption
-                      key={showFilter}
-                      value={showFilter}
-                      onClick={() => onShowFilterChanged?.(showFilter)}
-                      data-test={`show-${showFilter}`}
+                      key={filter}
+                      value={filter}
+                      onClick={() => onPreferenceChanged?.({ filter })}
+                      data-test={`show-${filter}`}
                     >
-                      {CollectionFilterName[showFilter]}
+                      {CollectionFilterName[filter]}
+                    </MenuItemOption>
+                  ))}
+                </MenuOptionGroup>
+
+                <MenuDivider />
+
+                <MenuOptionGroup
+                  value={collection?.grouping}
+                  title="Group by"
+                  type="radio"
+                >
+                  {CollectionGrouping.options.map((grouping) => (
+                    <MenuItemOption
+                      key={grouping}
+                      value={grouping}
+                      onClick={() => onPreferenceChanged?.({ grouping })}
+                      data-test={`grouping-${grouping}`}
+                    >
+                      {CollectionGroupingName[grouping]}
+                    </MenuItemOption>
+                  ))}
+                </MenuOptionGroup>
+
+                <MenuDivider />
+
+                <MenuOptionGroup
+                  value={collection?.sortBy}
+                  title="Sort by"
+                  type="radio"
+                >
+                  {CollectionSortBy.options.map((sortBy) => (
+                    <MenuItemOption
+                      key={sortBy}
+                      value={sortBy}
+                      onClick={() => onPreferenceChanged?.({ sortBy })}
+                      data-test={`sortBy-${sortBy}`}
+                    >
+                      {CollectionSortByName[sortBy]}
                     </MenuItemOption>
                   ))}
                 </MenuOptionGroup>