+++ packages/frontend/features/queryKeys.ts
@@ -1,4 +1,9 @@
-import { CollectionFilter, ID } from '@orpington-news/shared';
+import {
+  CollectionFilter,
+  CollectionGrouping,
+  CollectionSortBy,
+  ID,
+} from '@orpington-news/shared';
 
 export const collectionKeys = {
   base: ['collection'] as const,
@@ -7,8 +12,16 @@ export const collectionKeys = {
     [...collectionKeys.base, collectionId] as const,
   lists: (collectionId: ID) =>
     [...collectionKeys.base, collectionId, 'list'] as const,
-  list: (collectionId: ID, filter?: CollectionFilter) =>
-    [...collectionKeys.lists(collectionId), { filter }] as const,
+  list: (
+    collectionId: ID,
+    filter?: CollectionFilter,
+    grouping?: CollectionGrouping,
+    sortBy?: CollectionSortBy
+  ) =>
+    [
+      ...collectionKeys.lists(collectionId),
+      { filter, grouping, sortBy },
+    ] as const,
   detail: (collectionId: ID, itemId: ID) =>
     [...collectionKeys.base, collectionId, 'detail', { itemId }] as const,
 };